lib-index: mail_index_transaction_get_highest_modseq() - Fix handling MAIL_INDEX_MAIL_FLAG_UPDATE_MODSEQ
MAIL_INDEX_MAIL_FLAG_UPDATE_MODSEQ flag updates didn't calculate the
returned modseq correctly. This function wasn't used outside
--with-devel-checks though, but with it this fixes:

Panic: file mail-index-transaction.c: line 212 (mail_index_transaction_commit_real): assertion failed: (t->reset || expected_highest_modseq == log->head->sync_highest_modseq)

+++ src/lib-index/mail-index-transaction-export.c
@@ -579,7 +579,7 @@ transaction_flag_updates_have_non_internal(struct mail_index_transaction *t)
 	}
 
 	array_foreach(&t->updates, u) {
-		uint8_t changed_flags = u->add_flags | u->remove_flags;
+		uint16_t changed_flags = u->add_flags | u->remove_flags;
 
 		if ((changed_flags & ~internal_flags) != 0)
 			return TRUE;
